+++ dds_cli/gui_poc/pages/authentication/authentication_form.py
@@ -1,46 +1,16 @@
 from typing import Any
-
 from textual import events
 from textual.app import ComposeResult
 from textual.containers import Container
-from textual.widget import Widget
-from textual.widgets import Label
-
 from dds_cli.auth import Auth
-from dds_cli.gui_poc.components.dds_button import DDSButton, DDSFormButton
 from dds_cli.gui_poc.components.dds_form import DDSForm
 from dds_cli.gui_poc.components.dds_input import DDSInput
-from dds_cli.gui_poc.components.dds_modal import DDSModalConfirmation
-from dds_cli.gui_poc.components.dds_text_item import DDSTextSubtitle, DDSTextTitle
-from dds_cli.gui_poc.components.dds_container import DDSContainer
-from dds_cli.gui_poc.types.dds_severity_types import DDSSeverity
-
-class DDSLoginPage(DDSContainer):
-    """A widget for the Login page."""
-    def __init__(self, token_path: str, *args: Any, **kwargs: Any):
-        super().__init__(title="Login", *args, **kwargs)
-        self.token_path = token_path
-    
-    def compose(self) -> ComposeResult:
-        yield DDSTextTitle("Login")
-        yield DDSTextSubtitle("Please enter your credentials to authenticate with the DDS.")
-        yield DDSAuthForm(self.token_path)
+from dds_cli.gui_poc.components.dds_button import DDSFormButton
 
 
-class DDSReAuthenticatePage(DDSContainer):
-    """A widget for the Re-authenticate page."""
-    def __init__(self, token_path: str, *args: Any, **kwargs: Any):
-        super().__init__(title="Re-authenticate", *args, **kwargs)  
-        self.token_path = token_path
-
-    def compose(self) -> ComposeResult:
-        yield DDSTextTitle("Re-authenticate")
-        yield DDSTextSubtitle("Please enter your credentials to re-authenticate with the DDS.")
-        yield DDSAuthForm(self.token_path)
-
-
-class DDSAuthForm(Container):
+class AuthenticationForm(Container):
     """A widget for the authentication form."""
+
     def __init__(self, token_path: str, *args: Any, **kwargs: Any):
         super().__init__(*args, **kwargs)
         self.auth = None
@@ -48,17 +18,15 @@ class DDSAuthForm(Container):
 
     def compose(self) -> ComposeResult:
         with Container(id="dds-auth-form"):
-            yield DDSLoginFormFields()
+            yield LoginFormFields()
         with Container(id="dds-2fa-form"):
-            yield DDS2FAFormFields()
-
+            yield TwoFactorFormFields()
 
     def on_mount(self) -> None:
         """On mount, set the form to the login form."""
         self.query_one("#dds-auth-form").visible = True
         self.query_one("#dds-2fa-form").visible = False
-    
-    
+
     def on_button_pressed(self, event: events.Click) -> None:
         """Handle button presses."""
         if event.button.id == "send-2fa-code":
@@ -68,7 +36,6 @@ class DDSAuthForm(Container):
                 self.query_one("#dds-2fa-form").visible = True
         if event.button.id == "login":
             self.confirm_2factor_code()
-        
 
     def authenticate_user_credentials(self) -> None:
         """Authenticate the user credentials."""
@@ -87,45 +54,30 @@ class DDSAuthForm(Container):
         except Exception as e:
             self.notify(f"Error: {e}", severity="error")
             self.auth = None
-    
+
     def confirm_2factor_code(self) -> None:
         """Confirm the 2FA code."""
         code = self.query_one("#code").value
         try:
             self.auth.do_2factor(code)
             self.notify("Successfully logged in.")
-            #self.query_one(self.app.auth_status).update(True)
+            # self.query_one(self.app.auth_status).update(True)
         except Exception as e:
             self.notify(f"Error: {e}", severity="error")
 
-    
-class DDSLoginFormFields(DDSForm):
+
+class LoginFormFields(DDSForm):
     """A widget for the authentication form fields."""
+
     def compose(self) -> ComposeResult:
         yield DDSInput(placeholder="Username", id="username")
         yield DDSInput(placeholder="Password", id="password", password=True)
         yield DDSFormButton("Send 2FA code", id="send-2fa-code")
 
 
-class DDS2FAFormFields(DDSForm):
+class TwoFactorFormFields(DDSForm):
     """A widget for the 2FA form fields."""
+
     def compose(self) -> ComposeResult:
         yield DDSInput(placeholder="2FA code", id="code")
         yield DDSFormButton("Login", id="login")
-
-
-class DDSLogout(DDSModalConfirmation):
-    """A widget for the logout modal."""
-
-    def __init__(self, token_path: str, *args: Any, **kwargs: Any):
-        super().__init__(title="Logout", message="Are you sure you want to logout?", confirm_button_text="Logout", confirm_severity=DDSSeverity.WARNING, confirm_action=self.logout, *args, **kwargs)
-        self.token_path = token_path
-        #self.auth = Auth(token_path=self.token_path, authenticate=False)
-
-    def logout(self) -> None:
-        """Logout the user."""
-        self.app.auth.logout() #Do the logout action
-        self.app.compute_auth_status() #Compute the auth status
-        self.app.notify("Successfully logged out.")
-
-
