+++ dds_cli/gui_poc/pages/project_view.py
@@ -1,0 +1,67 @@
+"""DDS Project View Page"""
+
+from typing import Any
+
+from textual.app import ComposeResult
+from textual.containers import Vertical, Horizontal
+from textual.widget import Widget
+from dds_cli.gui_poc.pages.authentication.authentication import Authentication
+from dds_cli.gui_poc.pages.project_view_mode.project_list import ProjectList
+from dds_cli.gui_poc.pages.project_view_mode.project_information import ProjectInformation
+from dds_cli.gui_poc.pages.project_view_mode.project_content import ProjectContent
+from dds_cli.gui_poc.pages.project_view_mode.project_actions import ProjectActions
+from dds_cli.gui_poc.pages.project_view_mode.important_information import ImportantInformation
+
+
+class ProjectView(Widget):
+    """Project view page. Contains the project list, important information,
+    authentication menu, project content, project information, and project actions.
+    """
+
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)
+
+    DEFAULT_CSS = """
+    #left-container {
+        width: 25%;
+    }
+    #right-container {
+        width: 75%;
+    }
+    #project-list {
+        height: 40%;
+    }
+    #important-information {
+        height: 35%;
+    }
+    #auth-menu {
+        height: 25%;
+    }   
+    #top-container {
+        height: 50%;
+    }
+    #bottom-container {
+        height: 50%;
+    }   
+    #project-content {
+        width: 60%;
+    }
+    #project-information {
+        width: 40%;
+    }
+    """
+
+    def compose(self) -> ComposeResult:
+        with Horizontal():
+            with Vertical(id="left-container"):
+                yield ProjectList(title="Projects", id="project-list")
+                yield ImportantInformation(
+                    title="Important Information", id="important-information"
+                )
+                yield Authentication(title="Authentication", id="auth-menu")
+            with Vertical(id="right-container"):
+                with Horizontal(id="top-container"):
+                    yield ProjectContent(title="Project Content", id="project-content")
+                    yield ProjectInformation(title="Project Information", id="project-information")
+                with Horizontal(id="bottom-container"):
+                    yield ProjectActions(title="Project Actions", id="project-actions")
