+++ dds_cli/gui_poc/pages/project_view_mode/project_actions_tabs/download_data.py
@@ -21,8 +21,7 @@ class DownloadData(Widget):
 
     DEFAULT_CSS = """
     DownloadData {
-        border: round $secondary;
-        height: 10fr;
+        height: auto;
     }
 
     """
