+++ dds_cli/gui_poc/pages/project_view_mode/project_information.py
@@ -1,10 +1,12 @@
 """DDS Project Information Widget"""
 
-from typing import Any
 from textual.app import ComposeResult
-from textual.containers import Container, Vertical
-from textual.widgets import Label, Static
-from dds_cli.gui_poc.components.dds_container import DDSContainer, DDSContentContainer, DDSSpacedContainer
+from textual.widgets import Label
+from dds_cli.gui_poc.components.dds_container import (
+    DDSContainer,
+    DDSContentContainer,
+    DDSSpacedContainer,
+)
 from dds_cli.gui_poc.components.dds_key_pair_table import DDSKeyPairTable
 
 MOCK_PROJECT_INFORMATION = [
@@ -16,15 +18,13 @@ MOCK_PROJECT_INFORMATION = [
     ("Support contact", "Project Support Contact"),
 ]
 
-class DDSProjectInformation(DDSContainer):
-    """A widget for the project information."""
 
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
+class ProjectInformation(DDSContainer):
+    """A widget for the project information."""
 
     def compose(self) -> ComposeResult:
         with DDSSpacedContainer():
-            with DDSContentContainer():    
+            with DDSContentContainer():
                 yield Label("[b]Project Title:[/b] Test Title")
                 yield Label("[b]Project Description:[/b] Test Description")
             yield DDSKeyPairTable(MOCK_PROJECT_INFORMATION, id="project-information-table")
